+++ framework/web/UrlRule.php
@@ -8,8 +8,8 @@
 namespace yii\web;
 
 use Yii;
-use yii\base\Object;
 use yii\base\InvalidConfigException;
+use yii\base\Object;
 
 /**
  * UrlRule represents a rule used by [[UrlManager]] for parsing and generating URLs.
@@ -179,6 +179,14 @@ class UrlRule extends Object implements UrlRuleInterface
             $this->name = $this->pattern;
         }
 
+        $this->preparePattern();
+    }
+
+    /**
+     * Process [[$pattern]] on rule initialization.
+     */
+    private function preparePattern()
+    {
         $this->pattern = $this->trimSlashes($this->pattern);
         $this->route = trim($this->route, '/');
 
@@ -212,6 +220,18 @@ class UrlRule extends Object implements UrlRuleInterface
             }
         }
 
+        $this->translatePattern(true);
+    }
+
+    /**
+     * Prepares [[$pattern]] on rule initialization - replace parameter names by placeholders.
+     *
+     * @param bool $allowAppendSlash Defines position of slash in the param pattern in [[$pattern]].
+     * If `false` slash will be placed at the beginning of param pattern. If `true` slash position will be detected
+     * depending on non-optional pattern part.
+     */
+    private function translatePattern($allowAppendSlash)
+    {
         $tr = [
             '.' => '\\.',
             '*' => '\\*',
@@ -223,7 +243,10 @@ class UrlRule extends Object implements UrlRuleInterface
         ];
 
         $tr2 = [];
+        $requiredPatternPart = $this->pattern;
+        $oldOffset = 0;
         if (preg_match_all('/<([\w._-]+):?([^>]+)?>/', $this->pattern, $matches, PREG_OFFSET_CAPTURE | PREG_SET_ORDER)) {
+            $appendSlash = false;
             foreach ($matches as $match) {
                 $name = $match[1][0];
                 $pattern = isset($match[2][0]) ? $match[2][0] : '[^\/]+';
@@ -232,20 +255,46 @@ class UrlRule extends Object implements UrlRuleInterface
                 if (array_key_exists($name, $this->defaults)) {
                     $length = strlen($match[0][0]);
                     $offset = $match[0][1];
-                    if ($offset > 1 && $this->pattern[$offset - 1] === '/' && (!isset($this->pattern[$offset + $length]) || $this->pattern[$offset + $length] === '/')) {
+                    $requiredPatternPart = str_replace("/{$match[0][0]}/", '//', $requiredPatternPart);
+                    if (
+                        $allowAppendSlash
+                        && ($appendSlash || $offset === 1)
+                        && (($offset - $oldOffset) === 1)
+                        && isset($this->pattern[$offset + $length])
+                        && $this->pattern[$offset + $length] === '/'
+                        && isset($this->pattern[$offset + $length + 1])
+                    ) {
+                        // if pattern starts from optional params, put slash at the end of param pattern
+                        // @see https://github.com/yiisoft/yii2/issues/13086
+                        $appendSlash = true;
+                        $tr["<$name>/"] = "((?P<$placeholder>$pattern)/)?";
+                    } elseif (
+                        $offset > 1
+                        && $this->pattern[$offset - 1] === '/'
+                        && (!isset($this->pattern[$offset + $length]) || $this->pattern[$offset + $length] === '/')
+                    ) {
+                        $appendSlash = false;
                         $tr["/<$name>"] = "(/(?P<$placeholder>$pattern))?";
-                    } else {
-                        $tr["<$name>"] = "(?P<$placeholder>$pattern)?";
                     }
+                    $tr["<$name>"] = "(?P<$placeholder>$pattern)?";
+                    $oldOffset = $offset + $length;
                 } else {
+                    $appendSlash = false;
                     $tr["<$name>"] = "(?P<$placeholder>$pattern)";
                 }
+
                 if (isset($this->_routeParams[$name])) {
                     $tr2["<$name>"] = "(?P<$placeholder>$pattern)";
                 } else {
                     $this->_paramRules[$name] = $pattern === '[^\/]+' ? '' : "#^$pattern$#u";
                 }
             }
+        }
+
+        // we have only optional params in route - ensure slash position on param patterns
+        if ($allowAppendSlash && trim($requiredPatternPart, '/') === '') {
+            $this->translatePattern(false);
+            return;
         }
 
         $this->_template = preg_replace('/<([\w._-]+):?([^>]+)?>/', '<$1>', $this->pattern);
@@ -434,7 +483,7 @@ class UrlRule extends Object implements UrlRuleInterface
                 $url = substr($url, 0, $pos) . preg_replace('#/+#', '/', substr($url, $pos));
             }
         } elseif (strpos($url, '//') !== false) {
-            $url = preg_replace('#/+#', '/', $url);
+            $url = preg_replace('#/+#', '/', trim($url, '/'));
         }
 
         if ($url !== '') {
@@ -487,7 +536,8 @@ class UrlRule extends Object implements UrlRuleInterface
      * @param string $string
      * @return string
      */
-    private function trimSlashes($string) {
+    private function trimSlashes($string)
+    {
         if (strpos($string, '//') === 0) {
             return '//' . trim($string, '/');
         }